+++ source/type.spec.js
@@ -2,6 +2,10 @@ import { type as ramdaType } from 'ramda'
 
 import { type } from './type'
 
+test('with symbol', () => {
+  expect(type(Symbol())).toBe('Symbol')
+})
+
 test('with simple promise', () => {
   expect(type(Promise.resolve(1))).toBe('Promise')
 })