+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -2,6 +2,8 @@ WIP 6.4.0
 
 - Upgrade all `rollup` related dependencies
 
+- `R.type` supports `Symbol` just like *Ramda*.
+
 - Remove file extension in `main` property in `package.json` in order to allow `experimental-modules`. See also this Ramda's PR - https://github.com/ramda/ramda/pull/2678/files
 
 - Import `R.indexBy`/`R.when`/`R.zipObj`/`R.propEq`/`R.complement` changes from recent `@types/ramda` release.